requirements.txt
fastapi
uvicorn
python-jose[cryptography]
scikit-learn
joblib
python-dotenv
mlflow